Actually print command list in usage message
The xauth routine this was based on took three arguments:
a file pointer, a command name, and a prefix.

If a command name is given, prints help on that command,
else if it's null, prints help on all commands.

The iceauth version has only two arguments: file pointer & command name.
It was being called with file pointer & prefix, and since the prefix
matched no command names, no help was printed.

Since the way iceauth's help list is structured doesn't make it easy
to add a prefix to every line, skip indentation for now and just pass
NULL for the command argument.

Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>
Reviewed-by: Gaetan Nadon <memsize@videotron.ca>